feat(interaction): Moved handling of vegalite interaction and only calls for point up
